Harden auth, make the search indexes usable, and fix client type drift
Security. Argon2 was running on the event loop, and since it deliberately takes
about 100ms, one login stalled every other request in flight. It now runs on a
worker thread. Login also answered noticeably faster for an unknown address than
for a known one, which is a working way to find out who has an account no matter
how careful the error message is, so the unknown path now performs a matching
hash. Refresh rotation read the token, checked it, then wrote it back, so two
requests arriving together could both be issued a fresh pair and quietly defeat
the rotation; it now claims the row in one conditional update, and a test fires
two refreshes at once to prove only one wins.

Search. The trigram indexes could never be used, because similarity() in a WHERE
clause is not indexable and Postgres had to compute it for every row. Filtering
now uses the % operator with the threshold set per transaction, and the query
plan confirms a bitmap index scan. similarity() still orders the results, where
the cost is bounded by what % returned.

Client contract. Coordinates were going out as JSON strings, since the column is
Numeric and pydantic renders a Decimal as a string, while the app typed them as
numbers. They are numbers on the wire now and there is a test for it. A few
fields the server always sends were typed nullable on the client, which created
fallback branches that could never run.

Mobile. junk_ratio was read as a fraction in one place and guessed at in
another, so the formatter now follows the contract. The map only showed an error
when both of its queries failed, which rendered a partial outage as an innocent
empty state. The web target was declared and scripted but react-native-web is
not installed, so it is removed rather than left as a trap. Star sizing moved
into the design tokens, and there is now a committed .env.example explaining the
LAN address and the port.

+++ backend/app/models/restaurant.py
@@ -9,6 +9,7 @@ from __future__ import annotations
 
 import datetime as dt
 import uuid
+from decimal import Decimal
 
 from sqlalchemy import DateTime, ForeignKey, Index, Numeric, String, Uuid, func, text
 from sqlalchemy.orm import Mapped, mapped_column
@@ -27,8 +28,10 @@ class Restaurant(Base):
     )
     name: Mapped[str] = mapped_column(String(200), nullable=False)
     area: Mapped[str | None] = mapped_column(String(120))
-    latitude: Mapped[float | None] = mapped_column(Numeric(9, 6))
-    longitude: Mapped[float | None] = mapped_column(Numeric(9, 6))
+    # Numeric comes back as Decimal, not float. The API layer converts to a
+    # JSON number; the column stays exact.
+    latitude: Mapped[Decimal | None] = mapped_column(Numeric(9, 6))
+    longitude: Mapped[Decimal | None] = mapped_column(Numeric(9, 6))
     created_by: Mapped[uuid.UUID | None] = mapped_column(
         Uuid(as_uuid=True), ForeignKey("users.id", ondelete="SET NULL")
     )
